ykp_write_config: Only output flags if all bits match.
Do not output a false positive for a config/ticket flag if only
some bits match.

Without this patch, a configuration with the single cfgflag
STRONG_PW2 was shown as

  STRONG_PW2|OATH_FIXED_MODHEX2|OATH_FIXED_MODHEX

since there are some bits matching STRONG_PW2 (0x40) in
OATH_FIXED_MODHEX (0x50). With this patch, single cfgflag
STRONG_PW2 is shown as

  STRONG_PW2|OATH_FIXED_MODHEX2

OATH_FIXED_MODHEX2 is also shown as it shares the value of 0x40,
and the outputting is not context aware (yet).