ykp_write_config: Some modes now store four bytes key in uid.
OATH-HOTP and HMAC challenge response have 20 bytes key, but the
key field in the configuration is only 16 bytes. They therefor store
four more bytes in the uid field. The pretty-printer is now aware
of this, and will say "uid: n/a" when there is actually key data in
there, and show the complete 20 bytes key as "key:".