Add challenge-response tool (ykchalresp).
This required implementing multi-feature-report reads in ykcore.c.

It is kind of wrong to have these functions in libykpers, but that's
where they are at the moment. We might split some part of libykpers
out into a new library before releasing anything official. This is
to be considered work in progress.