Use supported mechanism to set ACL
According to Kay Sievers, one of the systemd / udev developers, in Red
Hat Bugzilla 811691, the udev-acl utility was never meant to be called
directly. In newer versions of systemd, the utility does not even exist
anymore.

The same goes for calling TAG directly in a .rules file.

Kay created a new device class in systemd / udev to label Yubico (and
other challenge / response devices) with, so that the ACL on the device
can be set in a supported way. This class is ID_SECURITY_TOKEN.

The file in this commit sets this class for Yubico devices (or at least,
for my Yubikey II, I don't know the idProduct for devices that I
do not own). This file should be used instead of the current
70-yubikey.rules file on systems that have a new enough systemd / udev
installed, starting with this commit in systemd master:
ff87b7e7486f954a18536bebbd83a81b7e7749b2