properly pad challenge-response with shorter than 64 byte challenge
For challenge response with short challenge (hmac-lt64 set) the
challenge is considered padded with a differet value than the last value
of the challenge. With nothing done this turns into 0 which means any 0
byte at the end of the challenge will be stripped. Here we make sure to
pad with something different than the last value (by xoring with 0xff).