Fix device access issues on macOS Catalina and later
When opening a connection to the USB HID device, always prefer the FIDO
interface over the keyboard one in the macOS platform layer. This is
because accessing keyboard and pointer HID devices is protected behind
explicit user dialogs starting in macOS Catalina. PAM modules are loaded
in various privileged system processes that do not trigger permission
prompts, so it is important that the support libraries prefer the FIDO
device instead, which does not require explicit user authorization.

Apple Legal ticket OSS-3019